Fixed user issue

+++ src/actions/index.ts
@@ -1,15 +1,18 @@
 "use server";
 
 import { prisma } from "@/utils/prisma";
+import { currentUser } from "@clerk/nextjs/server";
 import { revalidatePath } from "next/cache";
 
 export async function createTodo(formData: FormData) {
   try {
+    const user = await currentUser();
+    if (!user) return;
     const input = (formData.get("input") as string)?.trim();
     if (!input) return;
 
     await prisma.todo.create({
-      data: { title: input },
+      data: { title: input, userId: user.id },
     });
 
     revalidatePath("/");
@@ -20,9 +23,11 @@ export async function createTodo(formData: FormData) {
 
 export async function changeStatus(formData: FormData) {
   try {
+    const user = await currentUser();
+    if (!user) return;
     const inputId = formData.get("inputId") as string;
     const todo = await prisma.todo.findUnique({
-      where: { id: inputId },
+      where: { id: inputId, userId: user.id },
     });
 
     if (!todo) return;
@@ -40,13 +45,15 @@ export async function changeStatus(formData: FormData) {
 
 export async function editTodo(formData: FormData) {
   try {
+    const user = await currentUser();
+    if (!user) return;
     const newTitle = (formData.get("newTitle") as string)?.trim();
     const inputId = formData.get("inputId") as string;
 
     if (!newTitle) return; // Prevents empty updates
 
     await prisma.todo.update({
-      where: { id: inputId },
+      where: { id: inputId, userId: user.id },
       data: { title: newTitle },
     });
 
@@ -58,10 +65,12 @@ export async function editTodo(formData: FormData) {
 
 export async function deleteTodo(formData: FormData) {
   try {
+    const user = await currentUser();
+    if (!user) return;
     const inputId = formData.get("inputId") as string;
 
     await prisma.todo.delete({
-      where: { id: inputId },
+      where: { id: inputId, userId: user.id },
     });
 
     revalidatePath("/");
